backend/requirements.txt
fastapi
uvicorn[standard]
celery
redis
python-socketio
pymicrosoftautogen
langchain
chromadb-client
pydantic
python-dotenv
